# Conflicts:
#	VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/Configuration/vertimag-configuration.Prototype2.json

+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.DataLayer/Providers/Interfaces/ITorqueCurrentMeasurementsDataProvider.cs
@@ -7,9 +7,9 @@ namespace Ferretto.VW.MAS.DataLayer.Providers
     {
         #region Methods
 
-        TorqueCurrentMeasurementSession AddMeasurementSession(int? loadingUnitId, decimal? loadedNetWeight);
+        TorqueCurrentMeasurementSession AddMeasurementSession(int? loadingUnitId, double? loadedNetWeight);
 
-        TorqueCurrentSample AddSample(int sessionId, decimal value, System.DateTime timeStamp, System.DateTime requestTimeStamp);
+        TorqueCurrentSample AddSample(int sessionId, double value, System.DateTime timeStamp, System.DateTime requestTimeStamp);
 
         IEnumerable<TorqueCurrentMeasurementSession> GetAllMeasurementSessions();
